requirements.txt
addict==2.3.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==0.7.6
Django==3.1.1
django-chartjs==2.2.1
djangorestframework==3.11.1
elementpath==2.0.2
fastapi==0.61.1
flake8==3.8.4
graphviz==0.16
h11==0.10.0
idna==2.10
iso3166==1.0.1
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mt-940==4.23.0
packaging==20.4
pycodestyle==2.6.0
pycountry==20.7.3
pydantic==1.6.1
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.11
requests==2.24.0
rope==0.17.0
schwifty==2020.9.0
sepaxml==2.3.0
six==1.15.0
sqlparse==0.3.1
starlette==0.13.6
text-unidecode==1.3
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.12.1
webencodings==0.5.1
wrapt==1.12.1
xmlschema==1.2.3
